+++ src/irvine/oeis/a186/A186182.java
@@ -17,7 +17,7 @@ public class A186182 extends LambdaSequence {
   public A186182() {
     super(0, n -> {
       final int m = 4;
-      return (n == 0) ? Z.ONE : Rationals.SINGLETON.sum(1, n, k -> new Q(Binomial.binomial(n * (m + 1) - m * k - 1, n - k).multiply(k)).divide(n * m - (m - 1) * k)).num();
+      return (n == 0) ? Z.ONE : Rationals.SINGLETON.sum(1, n, k -> new Q(Binomial.binomial((long) n * (m + 1) - (long) m * k - 1, n - k).multiply(k)).divide(n * m - (long) (m - 1) * k)).num();
     });
   }
 }
