+++ src/irvine/oeis/a343/A343097.java
@@ -14,7 +14,7 @@ public class A343097 extends LambdaArray {
   public A343097() {
     super(0, 0, 0, -1, (n, k) -> {
       final long n2 = (long) n * n;
-      return Z.valueOf(k).pow(n2).add(Z.valueOf(k).pow((n2 + 3 * (n & 1)) / 4).multiply(2)).add(Z.valueOf(k).pow((n2 + (n % 2)) / 2)).add(Z.valueOf(k).pow(n * (n + 1) / 2).multiply(2)).add(Z.valueOf(k).pow(n * (n + (n & 1)) / 2).multiply(2)).divide(8);
+      return Z.valueOf(k).pow(n2).add(Z.valueOf(k).pow((n2 + 3 * (n & 1)) / 4).multiply(2)).add(Z.valueOf(k).pow((n2 + (n % 2)) / 2)).add(Z.valueOf(k).pow((long) n * (n + 1) / 2).multiply(2)).add(Z.valueOf(k).pow((long) n * (n + (n & 1)) / 2).multiply(2)).divide(8);
     });
   }
 }
