+++ src/irvine/oeis/a376/A376447.java
@@ -17,7 +17,7 @@ public class A376447 extends LambdaSequence {
   public A376447() {
     super(0, n -> {
       final int m = 7;
-      return (n == 0) ? Z.ONE : Rationals.SINGLETON.sum(1, n, k -> new Q(Binomial.binomial(n * (m + 1) - m * k - 1, n - k).multiply(k)).divide(n * m - (m - 1) * k)).num();
+      return (n == 0) ? Z.ONE : Rationals.SINGLETON.sum(1, n, k -> new Q(Binomial.binomial(n * (m + 1) - (long) m * k - 1, n - k).multiply(k)).divide((long) n * m - (long) (m - 1) * k)).num();
     });
   }
 }
